# Conflicts:
#	pkg/db/chat_log_model.go

+++ internal/interaction/ws_conn.go
@@ -101,7 +101,7 @@ func (u *WsConn) writeBinaryMsg(msg GeneralWsReq) (*websocket.Conn, error) {
 		if err != nil {
 			return nil, utils.Wrap(err, "SetWriteTimeout")
 		}
-		log.Debug("this msg length is :", len(buff.Bytes())/1024, "kb")
+		log.Debug("this msg length is :", float32(len(buff.Bytes()))/float32(1024), "kb")
 		if len(buff.Bytes()) > constant.MaxTotalMsgLen {
 			return nil, utils.Wrap(errors.New("msg too long"), utils.IntToString(len(buff.Bytes())))
 		}
